+++ src/img/CHANGES.md
@@ -1,8 +1,15 @@
 # imgadm changelog
 
+## 3.1.2
+
+- [OS-4097] support xz-compressed Docker images, e.g. learn/tutorial
+
+
 ## 3.1.1
 
+- [#412] support 'imgadm create' on non-cloned KVM VMs
 - [OS-3734] imgadm import fails with 'dataset already exists'
+
 
 ## 3.1.0